Move swagger into its own app file
Also include the API version in the docs
